% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_sigmark.R
\name{geom_sigmark}
\alias{geom_sigmark}
\title{Significance marker labels}
\usage{
geom_sigmark(colour_fill = NULL, colours, ...)
}
\arguments{
\item{colour_fill}{Either "colour" or "fill", depending on the geometric object used}

\item{colours}{A vector of colours to be passed into \verb{scale_*_manual()}}

\item{...}{Arguments to pass into \code{ggtext::geom_richtext()}}
}
\value{
A ggplot2 layer that can be added to a plot created with \code{ggplot2::ggplot()}.
}
\description{
A wrapper for \code{ggtext::geom_richtext()} designed to pass labels and their significance markers
}
\note{
The \code{prep_sigmark()} function returns a data frame with a \code{labels} variable. This should be used for the \code{label} aesthetic of \code{geom_sigmark()}.
Note that the function also includes a \verb{scale_*_manual} function call to ensure that the geom object is given the same colours as the significant markers.
Add \code{hjust = 0} for a horizontal bar chart or \code{vjust = 0} for a vertical bar chart.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(ggplot2)
mydf <- data.frame(smokers = c(rbinom(100, 1, .8),
                               rbinom(70, 1, 0.7),
                               rbinom(50, 1, 0.6)),
                   region = c(rep("A", 100), rep("B", 70), rep("C", 50)))
colour_vec <- c("red", "blue", "green", "orange")
names(colour_vec) <- LETTERS[1:4]

summ_df <- mydf |>
  count(smokers, region) |>
  filter(smokers == 1) |>
  mutate(perc = n / sum(n))

# Comparing subgroup to subgroup

test_res <- mydf |>
  pairwise_prop_test(smokers, region) |>
  filter(p_value < 0.05, smokers == 1) |>
  select(higher_group, lower_group)

summ_df |>
  prep_sigmark(test_res, region, perc, colour_vec, percent = TRUE, vs_rest = FALSE) |>
  ggplot(aes(x = perc, y = region, label = labels, fill = region)) +
  geom_col() +
  geom_sigmark("fill", colour_vec, hjust = 0) +
  ylim(c(0, 1))

# Comparing each subgroup to rest of sample

test_res2 <- mydf |>
  pairwise_prop_test(smokers, region, vs_rest = TRUE) |>
  filter(p_value < 0.05, smokers == 1) |>
  select(higher_group, lower_group)

summ_df |>
  prep_sigmark(test_res2, region, perc, colour_vec, percent = TRUE, vs_rest = TRUE) |>
  ggplot(aes(x = perc, y = region, label = labels, fill = region)) +
  geom_col() +
  geom_sigmark(hjust = 0) +
  ylim(c(0, 1))
}
