% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_prop_test_int.R
\name{pairwise_prop_test_int}
\alias{pairwise_prop_test_int}
\title{Internal function for Pairwise Proportional Tests}
\usage{
pairwise_prop_test_int(df, x, n, subgroups = NULL, ...)
}
\arguments{
\item{df}{A data frame}

\item{x}{A vector of "successes"}

\item{n}{A vector of "trials"}

\item{subgroups}{Optional, a vector of subgroup labels to append to the output}

\item{...}{Additional arguments passed on to pairwise.prop.test and prop.test}
}
\value{
A tibble listing each pairwise comparison and its corresponding p-value
}
\description{
Internal function to be used in pairwise_prop_test
}
