% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_sigmark.R
\name{prep_sigmark}
\alias{prep_sigmark}
\title{Prepare for geom_sigmark}
\usage{
prep_sigmark(df, test_df, subgroups, result, colours = NULL, percent, vs_rest)
}
\arguments{
\item{df}{A data frame with the summarized data to be charted}

\item{test_df}{The output of either \code{pairwise_t_test} or \code{pairwise_prop_test}. Note that you need to filter out that output to only include those comparisons to be shown as significant.}

\item{subgroups}{Grouping vector}

\item{result}{Result vector (either proportions or mean values)}

\item{colours}{A vector of colours associated with the subgroup levels. Not needed if \code{vs_rest} argument is set to \code{TRUE}.}

\item{percent}{A logical indicating whether the results are proportions or not}

\item{vs_rest}{A logical indicating whether the vs_rest argument in \verb{pairwise_*_test()} was set to TRUE or FALSE.}
}
\value{
A tibble ready for ggplot2 chart with a \code{geom_sigmark()} layer.
}
\description{
Helper function to prepare a summarized data frame for significance markers with geom_sigmark
}
\examples{
mydf <- data.frame(smokers = c(rbinom(100, 1, .8),
                               rbinom(70, 1, 0.7),
                               rbinom(50, 1, 0.6)),
                   region = c(rep("A", 100), rep("B", 70), rep("C", 50)))
test_res <- pairwise_prop_test(mydf, smokers, region)

colour_vec <- c("red", "blue", "green", "orange")
names(colour_vec) <- LETTERS[1:4]

library(dplyr)
mydf \%>\%
  count(smokers, region) \%>\%
  filter(smokers == 1) \%>\%
  mutate(perc = n / sum(n)) \%>\%
  prep_sigmark(test_res, region, perc, colour_vec, percent = TRUE, vs_rest = FALSE)
}
