% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_prop_test.R
\name{pairwise_prop_test}
\alias{pairwise_prop_test}
\title{Pairwise Proportional Tests}
\usage{
pairwise_prop_test(df, outcome, subgroups, vs_rest = FALSE, ...)
}
\arguments{
\item{df}{A data frame or tibble of raw observations}

\item{outcome}{Response vector}

\item{subgroups}{Grouping vector}

\item{vs_rest}{Logical indicating whether to test each level of a subgroup to the rest of the data}

\item{...}{Additional arguments passed on to pairwise.prop.test and prop.test}
}
\value{
A tibble with output from pairwise.prop.test
}
\description{
A tidy calculation of pairwise proportional comparisons between group levels with corrections for multiple testing
}
\examples{
mydf <- data.frame(smokers = c(rbinom(100, 1, 0.8),
                               rbinom(70, 1, 0.7),
                               rbinom(50, 1, 0.6)),
                   region = c(rep("A", 100), rep("B", 70), rep("C", 50)))
pairwise_prop_test(mydf, smokers, region)
pairwise_prop_test(mydf, smokers, region, vs_rest = TRUE)
}
