% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_t_test.R
\name{pairwise_t_test}
\alias{pairwise_t_test}
\title{Pairwise T Tests}
\usage{
pairwise_t_test(df, x, subgroups, ...)
}
\arguments{
\item{df}{A data frame or tibble}

\item{x}{Response vector}

\item{subgroups}{Grouping vector}

\item{...}{Additional arguments passed on to pairwise.t.test and t.test}
}
\value{
A tibble with output from pairwise.t.test
}
\description{
A tidy calculation of pairwise comparisons between group levels with corrections for multiple testing
}
\examples{
pairwise_t_test(iris, Sepal.Length, Species)
}
